+++ ddtrace/contrib/redis/asyncio_patch.py
@@ -1,8 +1,8 @@
 from ddtrace import config
-from ddtrace._trace.utils_redis import _run_redis_command_async
 from ddtrace._trace.utils_redis import _trace_redis_cmd
 from ddtrace._trace.utils_redis import _trace_redis_execute_async_cluster_pipeline
 from ddtrace._trace.utils_redis import _trace_redis_execute_pipeline
+from ddtrace.contrib.redis_utils import _run_redis_command_async
 
 from ...internal.utils.formats import stringify_cache_args
 from ...pin import Pin
