+++ ddtrace/contrib/yaaredis/patch.py
@@ -3,9 +3,9 @@ import os
 import yaaredis
 
 from ddtrace import config
-from ddtrace._trace.utils_redis import _run_redis_command_async
 from ddtrace._trace.utils_redis import _trace_redis_cmd
 from ddtrace._trace.utils_redis import _trace_redis_execute_pipeline
+from ddtrace.contrib.redis_utils import _run_redis_command_async
 from ddtrace.vendor import wrapt
 
 from ...internal.schema import schematize_service_name
